breakout cpu topology information to module
